requirements.txt
absl-py==0.15.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==2.15.8
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==3.2
django-cors-headers==4.4.0
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
exceptiongroup==1.2.0
flake8==5.0.4
flake8-broken-line==1.0.0
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-quotes==3.4.0
flake8-return==1.2.0
flatbuffers==1.12
frozenlist==1.4.1
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.1.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.1
iniconfig==2.0.0
isort==5.13.2
keras==2.15.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
lazy-object-proxy==1.10.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
mccabe==0.7.0
ml-dtypes==0.2.0
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.19.5
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.1
pefile==2023.2.7
pep8-naming==0.13.3
pillow==10.3.0
playsound==1.2.2
pluggy==0.13.1
protobuf==4.25.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydocstyle==6.3.0
pyflakes==2.5.0
pygame==2.5.2
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.10.0
PyMsgBox==1.0.9
PyOpenGL==3.1.7
pyperclip==1.8.2
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==0.1.30
PySocks==1.7.1
pyTelegramBotAPI==4.18.1
pytest==6.2.4
pytest-django==4.4.0
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
pytest-subtests==0.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytils==0.4.1
pytweening==1.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
telebot==0.0.5
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.19
Werkzeug==3.0.1
wrapt==1.12.1
yarl==1.9.4
zipp==3.17.0
